requirements.txt
azure-functions==1.11.2
requests==2.28.1
pydantic==1.10.2
